LMS/requirements.txt
Django==4.2.2
djangorestframework==3.15.2
psycopg2==2.9.10
Pillow==11.0.0
django-filter==0.5.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
django-cors-headers==4.6.0
requests==2.32.3